// src/shared/dataprovider/api/model/transactionSummaryData.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TransactionSummaryData { 
    Timestamp?: Date;
    TransactionNo?: number;
    OrderNumber? : string;
    ReferenceUserId?: string;
    Type?: string;
    MerchantName?: string;
    PMethod?: string;
    GrossAmount?: number;
    FixFee?: number;
    NetAmount?: number;
    Status?: number;
    ClientName?:string;
    VendorResponseTime?: string;
    HolderName? : string;
    AccountNumber? : string;
    ReferenceNumber? : string;
}